requirements.txt
annotated-types==0.7.0
anyio==4.12.1
boto3==1.42.43
botocore==1.42.43
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
exceptiongroup==1.3.1
filelock==3.20.3
fsspec==2026.1.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.4.0
idna==3.11
Jinja2==3.1.6
jmespath==1.1.0
joblib==1.5.3
jsonpatch==1.33
jsonpointer==3.0.0
langchain-core==1.2.9
langchain-text-splitters==1.1.0
langsmith==0.6.9
MarkupSafe==3.0.3
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.6
orjson==3.11.7
packaging==24.2
pinecone==8.0.0
pinecone-plugin-assistant==3.0.2
pinecone-plugin-interface==0.0.7
pydantic==2.12.5
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
PyYAML==6.0.3
regex==2026.1.15
requests==2.32.5
requests-toolbelt==1.0.0
s3transfer==0.16.0
safetensors==0.7.0
scikit-learn==1.7.2
scipy==1.15.3
sentence-transformers==5.2.2
shellingham==1.5.4
six==1.17.0
sympy==1.14.0
tenacity==9.1.3
threadpoolctl==3.6.0
tokenizers==0.22.2
torch==2.10.0
tqdm==4.67.3
transformers==5.1.0
typer-slim==0.21.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uuid_utils==0.14.0
xxhash==3.6.0
zstandard==0.25.0
PyMuPDF==1.25.3
sqlite-vec==0.1.6
python-dotenv==1.0.1
